+++ src/main/java/quizbank/repository/QuizRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import quizbank.dto.QuizDTO;
 import quizbank.model.Quiz;
 
-public interface QuizRepository extends JpaRepository<Quiz, Integer> {
+public interface QuizRepository extends JpaRepository<Quiz, Long> {
